+++ fosterParentDog/models/db/post_prefecture_db.go
@@ -1,0 +1,28 @@
+package db
+
+import (
+	"fmt"
+	"fpdapp/models/entity"
+)
+
+//*******************************************************************
+// [第1引数]の投稿IDで、譲渡可能都道府県を取得する
+//*******************************************************************
+func (db *Database) FindPostPrefectureList(postId string) ([]entity.PostPrefecture, error) {
+	var model []entity.PostPrefecture
+
+	//select post_prefecture.id, post_prefecture.post_prefecture_id from post_prefecture where post_id = 44;
+	err := db.connection.Table("post_prefectures").
+		Select("post_prefectures.id,"+
+			" post_prefectures.post_id,"+
+			" post_prefectures.post_prefecture_id").
+		Where("post_prefectures.post_id = ?", postId).
+		Scan(&model).
+		Error
+
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	return model, err
+}
